pyproject.toml
[project]
name = "spark-idea-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "numpy>=2.4.4",
    "pydantic-settings>=2.13.1",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.43.0",
]
